+++ OAuth/ResourceOwner/VkontakteResourceOwner.php
@@ -20,7 +20,7 @@ use Symfony\Component\OptionsResolver\Options;
  *
  * @author Adrov Igor <nucleartux@gmail.com>
  * @author Vladislav Vlastovskiy <me@vlastv.ru>
- * @author Alexander Latushkin <alex@skazo4neg.ru> 
+ * @author Alexander Latushkin <alex@skazo4neg.ru>
  */
 class VkontakteResourceOwner extends GenericOAuth2ResourceOwner
 {
@@ -31,6 +31,8 @@ class VkontakteResourceOwner extends GenericOAuth2ResourceOwner
         'identifier' => 'response.0.uid',
         'nickname'   => 'response.0.nickname',
         'profilepicture' => 'response.0.photo_50',
+        'firstname'  => 'response.0.first_name',
+        'lastname'   => 'response.0.last_name',
         'realname'   => array('response.0.last_name', 'response.0.first_name'),
         'email' => 'email'
     );
@@ -59,7 +61,7 @@ class VkontakteResourceOwner extends GenericOAuth2ResourceOwner
             $content['email'] = $accessToken['email'];
             $response->setResponse($content);
         }
-        
+
         return $response;
     }